add: 管理用seed実行API(/api/admin/seed)を追加
本番DBへ手元から直接接続しなくても、Basic認証付きのエンドポイントを
叩くだけで data/restaurants.csv の内容を本番DBに反映できるようにする。
CSVを更新してデプロイするたびに再利用できる運用導線。

- lib/seedRestaurants.js: seedロジックをCLI(prisma/seed.js)とAPIの
  共通モジュールに切り出し。IDシーケンス同期も含む
- app/api/admin/seed/route.js: GET/POSTでseedを実行し結果サマリを
  JSONで返す。middlewareにより/api/admin配下は常にBasic認証必須。
  Vercelのタイムアウト対策でmaxDuration=60を指定
- next.config.mjs: サーバーレス関数のバンドルにdata/restaurants.csvが
  含まれるようoutputFileTracingIncludesを設定

ローカルで認証なし401・認証ありで79店舗upsert+シーケンス同期の成功、
CLI版seedの動作継続、本番ビルドの成功を確認済み。
