+++ next.config.mjs
@@ -3,6 +3,12 @@ const nextConfig = {
   reactStrictMode: true,
   experimental: {
     esmExternals: 'loose',
+    // /api/admin/seed が実行時にCSVを読むため、サーバーレス関数の
+    // バンドルに data/ を含める(これがないとVercel上でファイルが見つからない)
+    outputFileTracingIncludes: {
+      '/api/admin/seed': ['./data/restaurants.csv'],
+      '/api/admin/seed/route': ['./data/restaurants.csv'],
+    },
   },
   images: {
     remotePatterns: [
